Functional analysis (SnpEff)
Functional analysis module is now adapted to work with normalized consensus vcf.
It is now run only once for consensus vcf only instead of vcf for three callers.

+++ modules/modeller.nf
@@ -9,10 +9,14 @@ process modeller {
 
     script:
     """
-    ls /home/SARS-CoV2/
-    find / -name modeller | grep -v opt
-    cp /home/SARS-CoV2/modeller/7dwz.pdb .
-    modpy.sh modeller_create_alignment.py ${target_fasta}
-    modpy.sh modeller_build_model.py alignment.pir
+    if [ -s ${target_fasta} ] ; then
+        cp /home/SARS-CoV2/modeller/7dwz.pdb .
+        modpy.sh modeller_create_alignment.py ${target_fasta}
+        modpy.sh modeller_build_model.py alignment.pir
+    else
+        echo "Empty fasta file. Skipping modeller."
+        touch alignment.pir
+        touch target.B99990001.pdb
+    fi
     """
 }
